15-iot-terraform/requirements.txt
azure-iot-device==2.14.0